tweaks and cleanup
