+++ .gitignore
@@ -8,7 +8,7 @@ appsettings.json
 appsettings.Development.json
 appsecrets.json
 images/
-[Ll]og*.txt
+[Ll]ogs/
 
 # User-specific files
 *.rsuser
