+++ src/MobileChat.MAUI/Pages/ChannelsPage.razor
@@ -1,9 +1,9 @@
 ﻿@page "/channels"
 
-@using JihadKhawaja.SignalR.Client.Chat.Interfaces
-@using JihadKhawaja.SignalR.Client.Chat.Models
+@using MobileChat.Client.Interfaces
 @using MobileChat.MAUI.Interfaces
 @using MobileChat.MAUI.Shared.Components
+@using MobileChat.Shared.Models
 @inject NavigationManager NavigationManager
 @inject ISaveFile SaveFileService
 @inject IChat ChatService
@@ -14,39 +14,43 @@
 }
 else if (!IsSignedIn)
 {
-    <SignInView />
+    <MudContainer MaxWidth="MaxWidth.Large" Class="my-4 pt-4">
+        <SignInView />
+    </MudContainer>
 }
 else
 {
-    <MudStack Class="mb-8" Row Justify="Justify.SpaceBetween" AlignItems="AlignItems.Center">
-        <MudText Typo="Typo.h3">Channels</MudText>
-        <MudButton OnClick="CreateChannel" Size="Size.Large" Variant="Variant.Filled" Color="Color.Primary" StartIcon="@Icons.Material.Filled.Add">Create Channel</MudButton>
-    </MudStack>
-    @if (CurrentUserChannels is null)
-    {
-        <MudText GutterBottom Typo="Typo.h6">No channels yet..</MudText>
-    }
-    else
-    {
-        <MudStack Justify="Justify.Center" AlignItems="AlignItems.Stretch" Spacing="0">
-            @foreach (var channel in CurrentUserChannels)
-            {
-                <MudPaper Class="mb-4 pa-2 pl-6 rounded-lg clickabe-paper" Elevation="2" onclick="@(() => OpenChannel(channel))">
-                    <MudGrid Justify="Justify.SpaceBetween">
-                        <MudItem Class="mt-3">
-                            <ChannelView Title=@channel.Title Users="@channel.Id.ToString()" />
-                        </MudItem>
-                        <MudItem>
-                            <MudStack Spacing="0">
-                                <MudButton>Edit</MudButton>
-                                <MudButton Color="Color.Error">Delete</MudButton>
-                            </MudStack>
-                        </MudItem>
-                    </MudGrid>
-                </MudPaper>
-            }
+    <MudContainer MaxWidth="MaxWidth.Large" Class="my-4 pt-4">
+        <MudStack Class="mb-8" Row Justify="Justify.SpaceBetween" AlignItems="AlignItems.Center">
+            <MudText Typo="Typo.h3">Channels</MudText>
+            <MudButton OnClick="CreateChannel" Size="Size.Large" Variant="Variant.Filled" Color="Color.Primary" StartIcon="@Icons.Material.Filled.Add">Create Channel</MudButton>
         </MudStack>
-    }
+        @if (CurrentUserChannels is null)
+        {
+            <MudText GutterBottom Typo="Typo.h6">No channels yet..</MudText>
+        }
+        else
+        {
+            <MudStack Justify="Justify.Center" AlignItems="AlignItems.Stretch" Spacing="0">
+                @foreach (var channel in CurrentUserChannels)
+                {
+                    <MudPaper Class="mb-4 pa-2 pl-6 rounded-lg clickabe-paper" Elevation="2" onclick="@(() => OpenChannel(channel))">
+                        <MudGrid Justify="Justify.SpaceBetween">
+                            <MudItem Class="mt-3">
+                                <ChannelView Title=@channel.Title Users="@channel.Id.ToString()" />
+                            </MudItem>
+                            <MudItem>
+                                <MudStack Spacing="0">
+                                    <MudButton>Edit</MudButton>
+                                    <MudButton Color="Color.Error">Delete</MudButton>
+                                </MudStack>
+                            </MudItem>
+                        </MudGrid>
+                    </MudPaper>
+                }
+            </MudStack>
+        }
+    </MudContainer>
 }
 
 @code{
