+++ src/MobileChat.MAUI/Pages/Index.razor
@@ -1,5 +1,7 @@
 ﻿@page "/"
 @inject NavigationManager NavigationManager
 
-<MudText Typo="Typo.h3" GutterBottom>MAUI Mobile Chat</MudText>
-<MudText Typo="Typo.h6" GutterBottom Color="Color.Secondary">Thank you for testing this out, Leave feedback when possible.</MudText>
+<MudContainer MaxWidth="MaxWidth.Large" Class="my-4 pt-4">
+    <MudText Typo="Typo.h3" GutterBottom>MAUI Mobile Chat</MudText>
+    <MudText Typo="Typo.h6" GutterBottom Color="Color.Secondary">Thank you for testing this out, Leave feedback when possible.</MudText>
+</MudContainer>
