+++ src/MobileChat.MAUI/Pages/SettingsPage.razor
@@ -1,5 +1,7 @@
 ﻿@page "/settings"
 @inject NavigationManager NavigationManager
 
-<MudText Typo="Typo.h3" GutterBottom>Settings</MudText>
-<MudText Typo="Typo.h6" GutterBottom Color="Color.Secondary">To be implemented</MudText>
+<MudContainer MaxWidth="MaxWidth.Large" Class="my-4 pt-4">
+    <MudText Typo="Typo.h3" GutterBottom>Settings</MudText>
+    <MudText Typo="Typo.h6" GutterBottom Color="Color.Secondary">To be implemented</MudText>
+</MudContainer>
