+++ src/MobileChat.MAUI/Shared/Components/SignInView.razor
@@ -1,11 +1,11 @@
 ﻿@page "/signin"
 
 @using FluentValidation
-@using JihadKhawaja.SignalR.Client.Chat.Interfaces
-@using JihadKhawaja.SignalR.Client.Chat.Models
 @using System.Text.RegularExpressions
+@using MobileChat.Client.Interfaces
 @using MobileChat.MAUI.Helpers
 @using MobileChat.MAUI.Interfaces
+@using MobileChat.Shared.Models
 @inject ISaveFile SaveFileService
 @inject IChat ChatService
 @inject NavigationManager NavigationManager
