+++ src/MobileChat.MAUI/Shared/Components/SignUpView.razor
@@ -1,10 +1,10 @@
 ﻿@page "/signup"
 
 @using FluentValidation
-@using JihadKhawaja.SignalR.Client.Chat.Interfaces
-@using JihadKhawaja.SignalR.Client.Chat.Models
 @using System.Text.RegularExpressions
-@using MobileChat.MAUI.Interfaces
+@using MobileChat.Client.Interfaces
+@using MobileChat.MAUI.Interfaces;
+@using MobileChat.Shared.Models
 @inject ISaveFile SaveFileService
 @inject IChat ChatService
 @inject NavigationManager NavigationManager
