+++ src/MobileChat.MAUI/Shared/MainLayout.razor
@@ -15,9 +15,7 @@
         <NavMenu />
     </MudDrawer>
     <MudMainContent>
-        <MudContainer MaxWidth="MaxWidth.Large" Class="my-4 pt-4">
-            @Body
-        </MudContainer>
+        @Body
     </MudMainContent>
 </MudLayout>
