+++ src/MobileChat.Server/Services/UserService.cs
@@ -1,8 +1,8 @@
-﻿using JihadKhawaja.SignalR.Server.Chat.Models;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using MobileChat.Server.Database;
 using MobileChat.Server.Helpers;
 using MobileChat.Server.Interfaces;
+using MobileChat.Shared.Models;
 
 namespace MobileChat.Server.Services
 {
@@ -144,7 +144,7 @@ namespace MobileChat.Server.Services
                     dbentry = context.Users.FirstOrDefault(x => x.Username == emailorusername && x.Password == password);
                 }
 
-                if(dbentry is not null)
+                if (dbentry is not null)
                 {
                     return Task.FromResult(true);
                 }
