+++ src/MobileChat.Shared/MobileChat.Shared.csproj
