+++ src/MobileChat.Shared/Models/Channel.cs
@@ -1,12 +1,13 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 
-namespace JihadKhawaja.SignalR.Client.Chat.Models
+namespace MobileChat.Shared.Models
 {
     public class Channel
     {
         [Key]
         public Guid Id { get; set; }
-        [Required]
+        [NotMapped]
         public string Title { get; set; }
         public DateTime? DateCreated { get; set; }
     }
